fix: add another ICS20 Authorization interface that will never change between implementations.
