+++ precompiles/authorization/IICS20Authorization.sol
@@ -91,3 +91,43 @@ interface IICS20Authorization {
 
 
 }
+
+
+/// @author Evmos Team
+/// @title Authorization Interface Standard
+/// @dev Events and functions that will not change between implementations
+interface IICS20AuthorizationStd {
+    /// @dev Emitted when an ICS-20 transfer authorization is granted.
+    /// @param grantee The address of the grantee.
+    /// @param granter The address of the granter.
+    /// @param allocations the Allocations authorized with this grant
+    event IBCTransferAuthorization(
+        address indexed grantee,
+        address indexed granter,
+        Allocation[] allocations
+    );
+
+    /// @dev This event is emitted when an granter revokes a grantee's allowance.
+    /// @param grantee The address of the grantee.
+    /// @param granter The address of the granter.
+    event RevokeIBCTransferAuthorization(
+        address indexed grantee,
+        address indexed granter
+    );
+
+    /// @dev Revokes IBC transfer authorization for a specific grantee.
+    /// @param grantee The address for which the transfer authorization will be revoked.
+    function revoke(address grantee) external returns (bool revoked);
+
+    /// @dev Returns the remaining number of tokens that a grantee smart contract
+    /// will be allowed to spend on behalf of granter through
+    /// IBC transfers. This is an empty by array.
+    /// @param grantee The address of the contract that is allowed to spend the granter's tokens.
+    /// @param granter The address of the account able to transfer the tokens.
+    /// @return allocations The remaining amounts allowed to spend for
+    /// corresponding source port and channel.
+    function allowance(
+        address grantee,
+        address granter
+    ) external view returns (Allocation[] memory allocations);
+}
