+++ precompiles/ics20/approve.go
@@ -4,13 +4,7 @@
 package ics20
 
 import (
-	"fmt"
-	"time"
-
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/cosmos/cosmos-sdk/x/authz"
-	authzkeeper "github.com/cosmos/cosmos-sdk/x/authz/keeper"
-	transfertypes "github.com/cosmos/ibc-go/v7/modules/apps/transfer/types"
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core/vm"
@@ -141,49 +135,4 @@ func (p Precompile) DecreaseAllowance(
 	}
 
 	return method.Outputs.Pack(true)
-}
-
-// AcceptGrant implements the ICS20 accept grant.
-func AcceptGrant(
-	ctx sdk.Context,
-	caller, origin common.Address,
-	msg *transfertypes.MsgTransfer,
-	authzAuthorization authz.Authorization,
-) (*authz.AcceptResponse, error) {
-	transferAuthz, ok := authzAuthorization.(*transfertypes.TransferAuthorization)
-	if !ok {
-		return nil, authz.ErrUnknownAuthorizationType
-	}
-
-	resp, err := transferAuthz.Accept(ctx, msg)
-	if err != nil {
-		return nil, err
-	}
-
-	if !resp.Accept {
-		return nil, fmt.Errorf(authorization.ErrAuthzNotAccepted, caller, origin)
-	}
-
-	return &resp, nil
-}
-
-// UpdateGrant implements the ICS20 authz update grant.
-func UpdateGrant(
-	ctx sdk.Context,
-	authzKeeper authzkeeper.Keeper,
-	grantee, origin common.Address,
-	expiration *time.Time,
-	resp *authz.AcceptResponse,
-) (err error) {
-	if resp.Delete {
-		err = authzKeeper.DeleteGrant(ctx, grantee.Bytes(), origin.Bytes(), authorization.TransferMsg)
-	} else if resp.Updated != nil {
-		err = authzKeeper.SaveGrant(ctx, grantee.Bytes(), origin.Bytes(), resp.Updated, expiration)
-	}
-
-	if err != nil {
-		return err
-	}
-
-	return nil
 }
