+++ precompiles/ics20/tx.go
@@ -69,7 +69,7 @@ func (p Precompile) Transfer(
 		}
 
 		// Accept the grant and return an error if the grant is not accepted
-		resp, err = AcceptGrant(ctx, contract.CallerAddress, origin, msg, auth)
+		resp, err = authorization.AcceptGrant(ctx, contract.CallerAddress, origin, msg, auth)
 		if err != nil {
 			return nil, err
 		}
@@ -83,7 +83,7 @@ func (p Precompile) Transfer(
 	// Update grant only if is needed
 	if contract.CallerAddress != origin {
 		// accepts and updates the grant adjusting the spending limit
-		if err = UpdateGrant(ctx, p.AuthzKeeper, contract.CallerAddress, origin, expiration, resp); err != nil {
+		if err = authorization.UpdateGrant(ctx, p.AuthzKeeper, contract.CallerAddress, origin, expiration, resp); err != nil {
 			return nil, err
 		}
 	}
